+++ modules/tinymce/src/plugins/help/test/ts/browser/DialogKeyboardNavTest.ts
@@ -1,5 +1,6 @@
 import { FocusTools, Keys, Mouse } from '@ephox/agar';
 import { before, describe, it } from '@ephox/bedrock-client';
+import { Obj } from '@ephox/katamari';
 import { SugarBody, SugarDocument } from '@ephox/sugar';
 import { TinyHooks, TinyUiActions } from '@ephox/wrap-mcagar';
 
@@ -72,6 +73,13 @@ describe('browser.tinymce.plugins.help.DialogKeyboardNavTest', () => {
     pressTabKey(editor);
     await pAssertFocusOnItem('Installed Plugins', 'div[role="document"]');
     pressTabKey(editor);
+    const installedPlugins = Obj.mapToArray(editor.plugins, (v, k) => k);
+    for (const installedPlugin of installedPlugins) {
+      await pAssertFocusOnItem(`Installed Plugins link:${installedPlugin}`, `a[href*="${installedPlugin}"]`);
+      pressTabKey(editor);
+    }
+    await pAssertFocusOnItem('Premium Plugins link:Learn more...', `a[href*="/pricing"]`);
+    pressTabKey(editor);
     await pAssertFocusOnItem('Close Button', '.tox-button:contains("Close")');
     pressTabKey(editor);
     await pAssertFocusOnItem('"x" Close Button', '.tox-button[title="Close"]');
@@ -86,6 +94,8 @@ describe('browser.tinymce.plugins.help.DialogKeyboardNavTest', () => {
     await pAssertFocusOnItem('Version Tab', '.tox-dialog__body-nav-item:contains("Version")');
     pressTabKey(editor);
     await pAssertFocusOnItem('TinyMCE Version', 'div[role="document"]');
+    pressTabKey(editor);
+    await pAssertFocusOnItem('Installed Plugins', `a[href*="tinymce"]`);
     pressTabKey(editor);
     await pAssertFocusOnItem('Close Button', '.tox-button:contains("Close")');
     pressTabKey(editor);